requirements.txt
Django==5.2.13
djangorestframework==3.17.1
cryptography==46.0.7
psycopg2-binary==2.9.12
